refactor(tests): DRY setup functions
