+++ test/express.js
@@ -7,6 +7,8 @@ var readTests = require('./integration/read')
 var updateTests = require('./integration/update')
 var deleteTests = require('./integration/delete')
 
+var db = require('./integration/setup')()
+
 function Express () {
   var app = express()
   app.use(bodyParser.json())
@@ -15,12 +17,36 @@ function Express () {
   return app
 }
 
+function setup (callback) {
+  db.initialize(function (err) {
+    if (err) {
+      return callback(err)
+    }
+
+    db.reset(callback)
+  })
+}
+
+function dismantle (app, server, callback) {
+  db.close(function (err) {
+    if (err) {
+      return callback(err)
+    }
+
+    if (app.close) {
+      return app.close(callback)
+    }
+
+    server.close(callback)
+  })
+}
+
 function runTests (createFn) {
   describe(createFn.name, function () {
-    createTests(createFn)
-    readTests(createFn)
-    updateTests(createFn)
-    deleteTests(createFn)
+    createTests(createFn, setup, dismantle)
+    readTests(createFn, setup, dismantle)
+    updateTests(createFn, setup, dismantle)
+    deleteTests(createFn, setup, dismantle)
   })
 }
